# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for the Dermify application",
        "title": "Dermify API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/modules/rf": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a radiofrequency module detail for a session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-rf"
                ],
                "summary": "Create RF module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RF module details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createRFModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.RFModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/modules/rf/{moduleId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an RF module detail by module ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-rf"
                ],
                "summary": "Get RF module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Module ID",
                        "name": "moduleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RFModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an RF module detail. Requires version for optimistic locking.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules-rf"
                ],
                "summary": "Update RF module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Module ID",
                        "name": "moduleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated RF details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateRFModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RFModuleDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apierrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH_INVALID_CREDENTIALS"
                },
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
...
        "handlers.RFModuleDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "depth": {
                    "type": "number",
                    "example": 2
                },
                "device_id": {
                    "type": "integer",
                    "example": 1
                },
                "energy_level": {
                    "type": "number",
                    "example": 30
                },
                "handpiece_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "module_id": {
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "type": "string"
                },
                "overlap": {
                    "type": "number",
                    "example": 10
                },
                "pulses_per_zone": {
                    "type": "integer",
                    "example": 3
                },
                "rf_mode": {
                    "type": "string",
                    "example": "bipolar"
                },
                "tip_type": {
                    "type": "string",
                    "example": "insulated"
                },
                "total_pulses": {
                    "type": "integer",
                    "example": 150
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer",
                    "example": 1
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.createRFModuleRequest": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "number"
                },
                "device_id": {
                    "type": "integer"
                },
                "energy_level": {
                    "type": "number"
                },
                "handpiece_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "overlap": {
                    "type": "number"
                },
                "pulses_per_zone": {
                    "type": "integer"
                },
                "rf_mode": {
                    "type": "string"
                },
                "tip_type": {
                    "type": "string"
                },
                "total_pulses": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.updateRFModuleRequest": {
            "type": "object",
            "properties": {
                "depth": {
                    "type": "number"
                },
                "device_id": {
                    "type": "integer"
                },
                "energy_level": {
                    "type": "number"
                },
                "handpiece_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "overlap": {
                    "type": "number"
                },
                "pulses_per_zone": {
                    "type": "integer"
                },
                "rf_mode": {
                    "type": "string"
                },
                "tip_type": {
                    "type": "string"
                },
                "total_pulses": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}